Change the spot creation strategy for CellContactDetector.
Now spots are directly created from the mask of contacts. Their position
are the mean position of the pixel in each binary object, and we deduce
its radius from the object size. Quality value stores the total
intensity of the contact image within the object. The threshold is used
to discard contacts with a small size.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>